fix(@clayui/data-provider): fixes error of not retrieving initial cache value when there are multiple instances for the same endpoint
This only fixes the case for the first request, it doesn't sync requests made by another instance in the component's life stream, for example when doing the refetch only the instance that made it is updated the others with the same endpoint are not and also for `loadMore `.

But if another instance does the refetch or `loadMore` and has the cache policy it will not make a new request when the cache policy is `cache-first`.

There isn't a real need for this yet, so it's not intentional for the DataProvider to magically do this behind the scenes, since we have a need for this we can plan for this feature in the future.