fix(@clayui/shared): fix bug of getting focused elements in document in menus
With the accessibility fix for menus with virtual cursor, which we added `aria-hidden` to all elements except the menu or the input for example, we can't tell which is the next focusable element if it's a button or input.

As FocusScope is only used for elements with React.Portal, it is very difficult to have a button or input with aria-hidden, the most recommendable for the button is to use disabled or aria-disabled and for input always use type=hidden, this it is safe to remove these two scenarios from our list.