feat(@clayui/core): add `expandable` API public to TreeView
Allows the developer to define according to some kind of rule whether the item should show the expand/collapse button or not, this
it's good for the use cases when it's loading an asynchronous item.

This commit also fixes the problem of clicking the expander button it didn't load the items asynchronously.