fix(@clayui/shared): fix error when losing focus in DropDown
DropDown only has the option to move the focus via arrow keys, when moving the focus to the first element when the menu opens, FocusScope registers the next element in the focusable document to recover the focus when the focus leaves the DropDown by pressing tab.

This wasn't happening because the Menu elements weren't focusable by FocusScope, so the focus was lost because we didn't know where the focus was within the menu.