refactor(@clayui/core): Implements a new way to initialize collection and cached rendering
Collection rendering lives in the component, but sometimes we have some problems because the API design depends on the item being rendered to get data from the item, for example the rendered text or the content.

It can also happen to always cause a rendering in the items because it is the last component of the component tree, moving the logic to the hook, we manage to create a cache system to only compute when necessary even if the component has many updates we will always return the collection cache and compute when child data changes or items change.

Another positive point is that by moving to the hook we managed to pre-initialize the tree in memory before the content is rendered in the DOM, keeping it in cache and rendering it in the DOM afterwards, a case for this is the Menu component that renders the content only when the Menu is visible, even if you open and close the menu, the content will always be the same as the one in the cache.

This refactoring only brings static and dynamic collection but we are not moving dynamic virtualization, there is still more work to be done on that.