feat(@clayui/shared): adds the possibility to have overlay stacked
There are use cases where you can have Overlay components stacked, for example DropDown with ColorPicker, Modal with any other Overlay. If the user clicks outside the overlay it will close all overlays, as well as any other interaction the overlay handles, like pressing esc.

This implementation only handle with events if the overlay is the last element of the stack, allowing to have an overlay stacked.