refactor(@clayui/core): refactors the collection virtualization implementation for the hook
This is just an initial refactoring but it still misses some features like the implementation of infinite scroll and is still not 100% adapted to the `useCollection` hook features, for example the layout is not assembled for all items only for the items visible in virtualization. We need to register all items in the layout even if they are not visible in virtualization, this has to work independently.