feat(@clayui/core): adds the step of extracting data from the collection separate from the rendering
This commit separates item data extraction from rendering because when the collection is enabled with virtualization we don't have visibility of all items only those visible in the UI which is a problem to integrate with the navigation and selection mechanism.

Virtualization does not work correctly if it is nested for example a DropDown with groups or TreeView for this reason it is disabled. Later we will move the TreeView and groups to a one-level virtualized list to avoid nesting elements.

This commit also changes some of the collection's APIs to return more information about the item, such as the position of the item in the list.

A point to note is that the layout of the list respects the filter if it is enabled so any change in children or items the list is reassembled again to avoid stale state.