feat(@clayui/nav): add new composition to Vertical Nav with collection API
This commit makes a lot of changes, first we are rewriting the VerticalNav component to use the Collection API and exposing a new public VerticalNav.Item component.

We are also adding a new `expandedKeys` API with the pattern of controlled and uncontrolled, this makes it easier to manage this state for items in the tree like implementing expand all or collapse.

We are also moving the implementation of all this to the `@clayui/core` package and the `@clayui/nav` package just keeps backwards compatibility with the old version of the component and possibility to use the new version when declaring the new composition.