chore(@clayui/core): adds new API to prevent list from being stale when having nested collections
This API only for very exclusive cases when the nested collection is created implicitly, when it is created implicitly the layout of the items is assembled according to the nested rendering in React.js and not created all at once recursively in the root of the collection, such as the cases of Picker with groups and DropDown where the Group that represents a collection is explicit in the component's composition.

For example in the case of an implicit collection, if the root component tries to get the item that belongs to another collection instance at some point in the rendering it may get an stale list because at that moment it has not yet rendered the nested collection in time to add the items to the layout.