fix(@clayui/multi-select): maintains backwards compatibility with the `menuRenderer` API but without the new features
We couldn't maintain compatibility with the `menuRenderer` API due to how the API was made, we couldn't have the visibility of the list or the items to implement all the new features to improve accessibility.

In this case we just keep the old behavior of how the menu worked without the accessibility improvements but only when the `menuRenderer` is declared. This prevents breaking change.