fix(@clayui/autocomplete): fix error when not validating value with items
This fixes the error when the initial value when rendering the component for the first time the `onChange` is called resetting the value for not being a selected value in the list.

Now this commit fixes this bug by validating the initial value with the list of items and also avoids updating the value on the first render.