fix(@clayui/core): Implements the use of cursor to reference unique items to move items in the tree
I'm changing the strategy when moving the items through the tree, using indexes is very fragile because removing an item in the tree can break all the indexes of the items when adding, this is similar when you have pagination of data using offset in the database that it can break when an item in the database is removed for example.

Instead we can use cursor which is a reference to the id of the item, in this implementation the `nodeByCursor` can be a bit slow for a very big list but these are edge cases with thousands or millions.

Using the cursor avoids the edge cases of every behavior we were trying to cover due to using indexes, being a tree it increases the amount of use cases and makes the code more complex. For now only `reorder` is using the cursor but we may soon switch to all methods using it too and removing indexes.