fix(@clayui/core): fix error when interacting with TreeView using SR JAWS
Using `role="application"` makes more complex TreeView interactions accessible to SRs or ATs that capture all keyboard input and only pass through to the browser those that are accessible according to the accessibility standard.

The `role="application"` we add in the root of the TreeView only for non-Apple platforms since mac or IOS SR doesn't have this problem. This allows the keyboard navigation for drag and drop to work properly and also the interactions on actions to work accordingly.