feat(@clayui/core): adds base implementation for rendering a table tree into an accessible list
This is the initial implementation to allow nested rows, the composition remains the same as rendering a simple table into a list.

When `nestedKey` is set we enable the treegrid pattern and list tree rendering. In this commit we have not yet added the expand mechanism that will be similar to TreeView.