fix(@clayui/core): fixes error when deselecting the indeterminate state for pre selected items in the first render
This behavior should only happen when items change, as in React.js, regardless of the first render it will invoke useEffect, in this scenario this ends up breaking the hydrate behavior of the first render for the selected items.

So we prevent this case from being computed on the first rendering and only being called when the items change.