feat(www): implements the new code block
Unfortunately, the sandpack is very slow and makes the page slow when there are many examples, so the benefit here is minimal.

I developed our own code block using renoun tokens that add types to our examples, also using react-live for the preview that allows us to quickly render our examples without depending on remote dependencies. We still maintain the button to open the example that has a preview in codesandbox, which is a big game changer.